backend/pyproject.toml
[project]
name = "ai-meerting-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "alembic>=1.17.2",
    "azure-identity>=1.25.1",
    "black>=25.11.0",
    "fastapi>=0.123.4",
    "ffmpeg-python>=0.2.0",
    "flake8>=7.3.0",
    "httpx>=0.28.1",
    "isort>=7.0.0",
    "openai>=2.8.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pydub>=0.25.1",
    "python-dotenv>=1.2.1",
    "python-magic>=0.4.27",
    "python-multipart>=0.0.20",
    "redis>=7.1.0",
    "reportlab>=4.0.0",
    "python-docx>=1.1.0",
    "sqlalchemy>=2.0.44",
    "tenacity>=9.1.2",
    "uvicorn>=0.38.0",
    "websockets>=15.0.1",
    "yt-dlp>=2025.11.12",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "pytest>=9.0.1",
    "watchdog>=6.0.0",
]
